+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+* [Issue #2] update images
+* [Issue #1] add detection of custom power plans
+
 ## 1.0.0
 
 * Initial version